Move the WireProtocol struct
