+++ cargo-flash/src/main.rs
@@ -18,7 +18,7 @@ use probe_rs::{
     probe::{
         daplink,
         debug_probe::{DebugProbe, DebugProbeError, DebugProbeType, MasterProbe},
-        protocol::WireProtocol,
+        WireProtocol,
         stlink,
     },
     flash::download::{download_file, Format},
