--- probe-rs/src/probe/protocol.rs
@@ -1,5 +1,0 @@
-#[derive(Copy, Clone, Debug)]
-pub enum WireProtocol {
-    Swd,
-    Jtag,
-}
